Move test to single file
